Se implemento los servicios para crear citas

+++ src/main/java/com/consultorio/dental_s/Repository/CitasRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CitasRepository extends JpaRepository<Citas, Integer> {
+public interface CitasRepository extends JpaRepository<Citas, Long> {
 }
